List comprehension example in 1-intro correction
I'm not a python expert but a starter. I use it at work sometimes for quick scripting. What I mean is that as a newbie I maybe wrong in my observation.

I was checking the 1 - Intro lesson and notice in Readme.md the following code 

    newlist = [x * 2 for number in sample]

and I receive the next error:
    Traceback (most recent call last):
      File "<stdin>", line 1, in <module>
    NameError: name 'x' is not defined

Reading the given documentation at http://docs.python.org/tutorial/datastructures.html#list-comprehensions, I concluded the code should be:
    newlist = [x * 2 for number in sample]
or 
    newlist = [number * 2 for number in sample]

I have the feeling the latest is way more descriptive, that's why I choose it.